Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@10829 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/renderer/opengl/ogl_g3dcom.cpp
@@ -5700,7 +5700,7 @@ void csGraphics3DOGLCommon::DrawTriangleMesh (G3DTriangleMesh& mesh)
       else
       {
 	// @@@ This is suspicious: inputtex is int but routine returns float?
-	inputtex = layer->GetStateFloat( csEffectStrings::texture_source );
+	inputtex = (int)layer->GetStateFloat( csEffectStrings::texture_source );
       }
       
       if( inputtex==-1 )
